requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
cloudinary==1.29.0
colorama==0.4.4
decorator==5.1.1
Django==4.0.3
djangorestframework==3.13.1
executing==0.8.3
install==1.3.5
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.28
pure-eval==0.2.2
Pygments==2.11.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
